requirements.txt
langchain
langchain-core
langchain-community
langgraph
langchain-google-genai
langchain-groq
langchain-mistralai
python-dotenv
faiss-cpu
tiktoken
fastapi
uvicorn
requests
streamlit
